style: add prettier
Adds prettier to the project.
Also add ".prettierignore" file for excluding "README.md" and
"CHANGELOG.md: and in this way, the README will not get formatted by
the prettier extension each time we try to save it. This is important
because it seems the prettier doesn't follow google's style guide for
markdown files.